+++ src/auto-recognition/init.mjs
@@ -1,14 +1,3 @@
-import { addNodeToRollBonus } from '../handlebars-handlers/add-bonus-to-item-sheet.mjs';
-import './extreme-mood-swings.mjs'
+import './extreme-mood-swings.mjs';
+import './improved-critical.mjs';
 import './weapon-finesse.mjs';
-
-Hooks.on(
-    'renderItemSheet',
-    (
-        /** @type {ItemSheetPF} */ { isEditable, item },
-        /** @type {[HTMLElement]} */[html],
-        /** @type {unknown} */ _data
-    ) => {
-        addNodeToRollBonus(html, null, item, isEditable);
-    }
-);